Allow caller to pass in call config
